test_framework/requirements.txt
pytest==7.4.3
pytest-mock==3.12.0
pytest-xdist==3.5.0
requests==2.31.0
allure-pytest==2.13.2
python-dotenv==1.0.0
tenacity==8.2.3
pydantic==2.5.0